feat: infer standard methods without singular and plural
Many services adhering to AIP are not setting these two fields on the
resource descriptors, notably most of googleapis. This means that is not
possible to only infer the name of the method and assume that it is
correct.

Instead each method descriptor is inspected to see whether it is a
standard method, and if so which resource it refers to.