fix(all): generate top level skip check
Previously, code generated outside individual test cases would still be
run, even if the entire method was skipped.

For example with `Skip: []string{"Get"}` would mean that all test cases
for Get would be ignored, but the setup code would still be run.

This commit adds a top level skip check, which prevents this code to be
run if every test case is ignored.