fix: allow safe creation of prerequired resources by not providing invalid parent
In order to create a certain resource, other resources might need to be in place first.
The TestSuiteConfig's Create method should allow the developer to satisfy this precondition.
If the precondition requires creating resources from the same parent as the target resource,
we should not pass them an invalid parent. An invalid parent could fail the creation of the
prerequired resources, thus preventing the precondition for the test to be reached.