feat: allow providing a IDGenerator function
For services with user-settable IDs, there might be requirements on how
the format of that ID should look. Providing this function, i.e. it is
non-nil, will call it and fill in the returned ID during creation steps.

It is expected (and documented) that the function should return unique
IDs. It is up to the user to fulfill this requirement. One provided
example is to use a counter and include that in the ID.

This feature makes it possible to use this test suite for APIs where
user-settable ID is required, which wasn't possible before.