feat: bump googleapis ref and change testconfig name
Since googleapis don't have any version number we have to depend on a specific commit. This
leads to us having to update the ref after 1_000 commits.

Since the last commit, the featurestore service have been added, which imports the same
resources as other services. This lead to conflict in the test files. To fix this we now
prefix the service name to the test-config structs.